Integrate bus bot into main script
