+++ scal_full_integrated.py
@@ -64,7 +64,7 @@ Fully-integrated Smart Frame + Telegram bot with Google OAuth routes
 import os, json, time, secrets, threading, collections, re, socket, fcntl, xml.etree.ElementTree as ET
 from pathlib import Path
 from datetime import datetime, timezone, timedelta, date
-from typing import List, Optional, Tuple
+from typing import Any, Dict, List, Optional, Tuple
 from urllib.parse import quote
 
 import yaml
@@ -501,6 +501,270 @@ def fetch_bus():
     items.sort(key=lambda x: x["min1"] if x["min1"] is not None else 9999)
     items = items[:14]
     return {"city_code": city, "stop_name": stop_name, "node_id": node, "items": items}
+
+
+# === [SECTION: Standalone Bus Arrival Telegram Bot (PTB)] ====================
+try:
+    from telegram import (
+        Update,
+        InlineKeyboardButton,
+        InlineKeyboardMarkup,
+        ForceReply,
+    )
+    from telegram.ext import (
+        Application,
+        CallbackQueryHandler,
+        CommandHandler,
+        ContextTypes,
+        MessageHandler,
+        filters,
+    )
+except ModuleNotFoundError:  # pragma: no cover - 테스트 환경용 더미
+    Update = InlineKeyboardButton = InlineKeyboardMarkup = ForceReply = object  # type: ignore
+    Application = CallbackQueryHandler = CommandHandler = ContextTypes = MessageHandler = object  # type: ignore
+
+    class filters:  # type: ignore
+        TEXT = COMMAND = None
+
+BB_BOT_TOKEN = CFG["telegram"]["bot_token"]
+BB_ALLOWED_USER_IDS = set(CFG["telegram"].get("allowed_user_ids", []))
+BB_TAGO_SERVICE_KEY = os.environ.get("TAGO_API_KEY", CFG.get("bus", {}).get("key", "")).strip()
+BB_DEFAULT_CITY = CFG.get("bus", {}).get("city_code", "")
+BB_DEFAULT_NODE = CFG.get("bus", {}).get("node_id", "")
+
+BB_USER_STATE: Dict[int, Dict[str, Any]] = {}
+BB_CITY_CACHE: List[Tuple[str, str]] = []
+bb_log = logging.getLogger("busbot")
+
+
+def bb_ensure_user_state(uid: int) -> Dict[str, Any]:
+    if uid not in BB_USER_STATE:
+        BB_USER_STATE[uid] = {
+            "city_code": BB_DEFAULT_CITY,
+            "node_id": BB_DEFAULT_NODE,
+            "key": BB_TAGO_SERVICE_KEY,
+            "awaiting": None,
+            "stop_results": [],
+        }
+    return BB_USER_STATE[uid]
+
+
+def bb_check_auth(update: Update) -> bool:
+    u = update.effective_user
+    return bool(u and u.id in BB_ALLOWED_USER_IDS)
+
+
+def bb_extract_arg(text: str) -> str:
+    parts = text.strip().split(maxsplit=2)
+    if len(parts) == 2:
+        return parts[1]
+    if len(parts) >= 3:
+        return parts[2]
+    return ""
+
+
+def bb_extract_arg2(text: str) -> List[str]:
+    parts = text.strip().split()
+    return parts[2:] if len(parts) >= 3 else []
+
+
+def bb_is_tago_node_id(text: str) -> bool:
+    """ICB164000104 형태의 TAGO nodeId인지 검사"""
+    return bool(re.fullmatch(r"[A-Z]{3}\d{7,}", text.strip()))
+
+
+def bb_paginate(items: List[Any], page: int, per_page: int = 10) -> List[Any]:
+    start = page * per_page
+    end = start + per_page
+    return items[start:end]
+
+
+def bb_tago_get_city_list(service_key: str) -> List[Tuple[str, str]]:
+    """(cityName, cityCode) 리스트 반환"""
+    if BB_CITY_CACHE:
+        return BB_CITY_CACHE
+    url = (
+        "http://apis.data.go.kr/1613000/BusRouteInfoInqireService/getCtyCodeList"
+        f"?serviceKey={quote(service_key)}"
+    )
+    r = requests.get(url, timeout=7)
+    r.raise_for_status()
+    root = ET.fromstring(r.text)
+    for it in root.iter("item"):
+        name = _pick_text(it, "cityname") or _pick_text(it, "cityName")
+        code = _pick_text(it, "citycode") or _pick_text(it, "cityCode")
+        if name and code:
+            BB_CITY_CACHE.append((name, code))
+    return BB_CITY_CACHE
+
+
+def bb_tago_search_stops(city_code: str, keyword: str, service_key: str) -> List[Tuple[str, str, str]]:
+    """(정류소명, 정류소번호, nodeId) 리스트 반환"""
+    url = (
+        "http://apis.data.go.kr/1613000/BusSttnInfoInqireService/getSttnList"
+        f"?serviceKey={quote(service_key)}&cityCode={quote(str(city_code))}&nodeNm={quote(keyword)}"
+    )
+    r = requests.get(url, timeout=7)
+    r.raise_for_status()
+    root = ET.fromstring(r.text)
+    stops: List[Tuple[str, str, str]] = []
+    for it in root.iter("item"):
+        name = _pick_text(it, "nodenm") or _pick_text(it, "nodeNm")
+        ars = _pick_text(it, "arsno") or _pick_text(it, "arsNo")
+        node = _pick_text(it, "nodeid") or _pick_text(it, "nodeId")
+        if name and node:
+            stops.append((name, ars, node))
+    return stops
+
+
+def bb_build_city_keyboard(cities: List[Tuple[str, str]], page: int = 0) -> InlineKeyboardMarkup:
+    per = 10
+    items = bb_paginate(cities, page, per)
+    buttons = [
+        [InlineKeyboardButton(f"{name}", callback_data=f"city:{code}")]
+        for name, code in items
+    ]
+    nav = []
+    if page > 0:
+        nav.append(InlineKeyboardButton("⬅️", callback_data=f"citypage:{page-1}"))
+    if (page + 1) * per < len(cities):
+        nav.append(InlineKeyboardButton("➡️", callback_data=f"citypage:{page+1}"))
+    if nav:
+        buttons.append(nav)
+    return InlineKeyboardMarkup(buttons)
+
+
+def bb_build_stop_keyboard(stops: List[Tuple[str, str, str]], page: int = 0) -> InlineKeyboardMarkup:
+    per = 10
+    items = bb_paginate(stops, page, per)
+    buttons = [
+        [InlineKeyboardButton(f"{name} {ars}", callback_data=f"stop:{node}")]
+        for name, ars, node in items
+    ]
+    nav = []
+    if page > 0:
+        nav.append(InlineKeyboardButton("⬅️", callback_data=f"stoppage:{page-1}"))
+    if (page + 1) * per < len(stops):
+        nav.append(InlineKeyboardButton("➡️", callback_data=f"stoppage:{page+1}"))
+    if nav:
+        buttons.append(nav)
+    return InlineKeyboardMarkup(buttons)
+
+
+async def bb_start(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    if not bb_check_auth(update):
+        return
+    st = bb_ensure_user_state(update.effective_user.id)
+    await update.message.reply_text(
+        "🚌 버스 도착정보 봇\n"
+        "- /bus\n"
+        "- /stop (정류소 변경)\n"
+        "- /set key <서비스키>\n\n"
+        f"현재설정: city={st['city_code']} / node={st['node_id']} / key={'등록됨' if st['key'] else '미등록'}"
+    )
+
+
+async def bb_cmd_bus(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    if not bb_check_auth(update):
+        return
+    st = bb_ensure_user_state(update.effective_user.id)
+    city, node, key = st["city_code"], st["node_id"], st["key"]
+    await update.message.reply_text(f"⏳ 조회 중… (city={city}, node={node})")
+    stop_name, lines = tago_get_arrivals(city, node, key)
+    await update.message.reply_text("\n".join(lines) if lines else "정보가 없습니다.")
+
+
+async def bb_cmd_set(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    if not bb_check_auth(update):
+        return
+    st = bb_ensure_user_state(update.effective_user.id)
+    text = (update.message.text or "").strip()
+    if text.lower().startswith("/set key"):
+        arg = bb_extract_arg(text)
+        if not arg:
+            await update.message.reply_text("사용법: /set key <서비스키>")
+            return
+        st["key"] = arg.strip()
+        await update.message.reply_text("✔ 서비스키 등록 완료")
+    else:
+        await update.message.reply_text("사용법: /set key <서비스키>")
+
+
+async def bb_cmd_stop(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    if not bb_check_auth(update):
+        return
+    st = bb_ensure_user_state(update.effective_user.id)
+    cities = bb_tago_get_city_list(st["key"])
+    await update.message.reply_text(
+        "도시를 선택하세요", reply_markup=bb_build_city_keyboard(cities, 0)
+    )
+
+
+async def bb_on_callback(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    if not bb_check_auth(update):
+        return
+    query = update.callback_query
+    await query.answer()
+    data = query.data or ""
+    uid = update.effective_user.id
+    st = bb_ensure_user_state(uid)
+
+    if data.startswith("citypage:"):
+        page = int(data.split(":", 1)[1])
+        cities = bb_tago_get_city_list(st["key"])
+        await query.edit_message_reply_markup(bb_build_city_keyboard(cities, page))
+        return
+    if data.startswith("city:"):
+        code = data.split(":", 1)[1]
+        st["city_code"] = code
+        st["awaiting"] = "keyword"
+        await query.message.reply_text(
+            "정류소명을 입력하세요", reply_markup=ForceReply(selective=True)
+        )
+        return
+    if data.startswith("stoppage:"):
+        page = int(data.split(":", 1)[1])
+        stops = st.get("stop_results", [])
+        await query.edit_message_reply_markup(bb_build_stop_keyboard(stops, page))
+        return
+    if data.startswith("stop:"):
+        node = data.split(":", 1)[1]
+        st["node_id"] = node
+        st["awaiting"] = None
+        await query.message.reply_text(f"✔ 정류소 등록 완료: {node}")
+        return
+
+
+async def bb_on_message(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    if not bb_check_auth(update):
+        return
+    uid = update.effective_user.id
+    st = bb_ensure_user_state(uid)
+    if st.get("awaiting") == "keyword":
+        kw = (update.message.text or "").strip()
+        if bb_is_tago_node_id(kw):
+            st["node_id"] = kw
+            st["awaiting"] = None
+            await update.message.reply_text(f"✔ 정류소 등록 완료: {kw}")
+            return
+        stops = bb_tago_search_stops(st["city_code"], kw, st["key"])
+        st["stop_results"] = stops
+        st["awaiting"] = "stop"
+        await update.message.reply_text(
+            "정류소를 선택하세요", reply_markup=bb_build_stop_keyboard(stops, 0)
+        )
+
+
+def run_bus_bot():
+    app = Application.builder().token(BB_BOT_TOKEN).build()
+    app.add_handler(CommandHandler("start", bb_start))
+    app.add_handler(CommandHandler("bus", bb_cmd_bus))
+    app.add_handler(CommandHandler("set", bb_cmd_set))
+    app.add_handler(CommandHandler("stop", bb_cmd_stop))
+    app.add_handler(CallbackQueryHandler(bb_on_callback))
+    app.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, bb_on_message))
+    bb_log.info("Bus bot started.")
+    app.run_polling()
 
 
 # === [SECTION: Photo file listing for board background] ======================
